refactor(sdk): extract hex encoding utilities into separate module
